requirements.txt
vllm
huggingface_hub
websocket-client
pyautogui
pywinauto
pywin32
psutil
keyboard
playwright
httpx
trafilatura
readability-lxml
opencv-python
pillow
pytesseract
faster-whisper
edge-tts
sounddevice
chromadb
sentence-transformers
python-magic-bin
pefile
lief
yara-python
construct
capstone
pydantic
pydantic-settings
toml
websockets
watchdog
openai
